Updating to CM6 stable (rough)
