+++ components/CodeEditorOld/index.js
@@ -1,0 +1,3 @@
+import CodeEditor from "./CodeEditor";
+
+export default CodeEditor;
