+++ components/CodeMirror6Instance/index.js
@@ -1,0 +1,1 @@
+export { default } from "./CodeMirror6Instance";
