+++ components/EditorSettings/EditorSettings.module.scss
@@ -13,6 +13,15 @@
   }
 }
 
+.option {
+  &[data-implemented="false"] {
+    opacity: 0.5;
+    &:hover {
+      opacity: 0.8;
+    }
+  }
+}
+
 .notes {
   font-size: 0.7rem;
   margin: 0.5rem 0;
