+++ package.json
@@ -10,29 +10,35 @@
     "lint": "next lint",
     "analyze": "ANALYZE=true yarn build"
   },
+  "oldDependencies": {
+    "//": "These are no longer applicable to the public release of CodeMirror 6. They've been merged into some of the base packages.",
+    "@codemirror/history": "^6.0.0",
+    "@codemirror/highlight": "^6.0.0",
+    "@codemirror/gutter": "^6.0.0",
+    "@codemirror/fold": "^6.0.0",
+    "@codemirror/comment": "^6.0.0",
+    "@codemirror/closebrackets": "^6.0.0",
+    "@codemirror/rectangular-selection": "^6.0.0"
+  },
   "dependencies": {
-    "@codemirror/autocomplete": "^0.18.8",
-    "@codemirror/closebrackets": "^0.18.0",
-    "@codemirror/commands": "^0.18.3",
-    "@codemirror/comment": "^0.18.1",
-    "@codemirror/fold": "^0.18.1",
-    "@codemirror/gutter": "^0.18.4",
-    "@codemirror/highlight": "^0.18.4",
-    "@codemirror/history": "^0.18.1",
-    "@codemirror/lang-css": "^0.18.0",
-    "@codemirror/lang-html": "^0.18.1",
-    "@codemirror/lang-javascript": "^0.18.0",
-    "@codemirror/lang-markdown": "^0.18.4",
-    "@codemirror/language": "^0.18.2",
-    "@codemirror/lint": "^0.18.4",
-    "@codemirror/rectangular-selection": "^0.18.0",
-    "@codemirror/search": "^0.18.4",
-    "@codemirror/state": "^0.18.7",
-    "@codemirror/view": "^0.18.12",
+    "@codemirror/autocomplete": "^6.0.0",
+    "@codemirror/commands": "^6.0.0",
+    "@codemirror/lang-css": "^6.0.0",
+    "@codemirror/lang-html": "^6.0.0",
+    "@codemirror/lang-javascript": "^6.0.0",
+    "@codemirror/lang-markdown": "^6.0.0",
+    "@codemirror/language": "^6.0.0",
+    "@codemirror/legacy-modes": "^6.0.0",
+    "@codemirror/lint": "^6.0.0",
+    "@codemirror/search": "^6.0.0",
+    "@codemirror/state": "^6.0.0",
+    "@codemirror/view": "^6.0.0",
+    "@emmetio/codemirror6-plugin": "^0.1.1",
     "@netlify/plugin-nextjs": "3.5.0",
     "@next/bundle-analyzer": "^11.0.1",
     "classnames": "^2.3.1",
-    "next": "11.0.1",
+    "codemirror": "^6.0.0",
+    "next": "^12.1.4",
     "prettier": "^2.3.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
